requirements.txt
pdfplumber
python-docx
sentence-transformers
spacy
pandas
torch
scikit-learn
google-generativeai
